Break out idGenerationHelpers to own module
